+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `non-interactive` flag to `flash` subcommand (#737)
 - Add `no-reset` flag to `monitor` subcommands (#737)
 - Add an environment variable to set monitoring baudrate (`MONITOR_BAUD`) (#737)
+Add list-ports command to list available serial ports.
+- Add list-ports command to list available serial ports.
 
 ### Changed
 - Split the baudrate for connecting and monitorinig in `flash` subcommand (#737)